+++ backend/src/test/java/com/example/demo/controllers/UsersControllerTest.java
@@ -1,10 +1,15 @@
 package com.example.demo.controllers;
 
 import com.example.demo.builders.ClientUserBuilder;
+import com.example.demo.builders.StoreAdminBuilder;
+import com.example.demo.helpers.StoreTestHelper;
 import com.example.demo.model.exceptions.NotFoundUserException;
+import com.example.demo.model.store.Store;
+import com.example.demo.model.user.StoreAdminUser;
 import com.example.demo.services.users.UserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.example.demo.model.user.ClientUser;
+import com.jayway.jsonpath.JsonPath;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +23,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
 import static org.hamcrest.Matchers.is;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
@@ -29,6 +35,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @AutoConfigureMockMvc
 @ActiveProfiles("test")
 public class UsersControllerTest {
+
+    private StoreTestHelper storeTestHelper = new StoreTestHelper();
 
     @MockBean
     UserService userServiceMock;
@@ -76,5 +84,31 @@ public class UsersControllerTest {
                 .andReturn();
         String response = mvcResult.getResponse().getContentAsString();
 
+    }
+
+    @Test
+    public void addingAValidStoreAdminReturnsTheStoreAdminAnd200Status() throws Exception {
+        StoreAdminUser aStoreAdmin = StoreAdminBuilder.aStoreAdmin().build();
+        Store aStore = aStoreAdmin.store();
+        when(userServiceMock.addStoreAdmin(any())).thenReturn(aStoreAdmin);
+
+        String content = objectMapper.writeValueAsString(aStoreAdmin);
+        MvcResult mvcResult = mockMvc.perform(post("/storeAdmin")
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(content))
+                .andExpect(status().isOk())
+                .andReturn();
+
+        String response = mvcResult.getResponse().getContentAsString();
+        assertEquals(JsonPath.parse(response).read("username"), aStoreAdmin.username());
+        assertEquals(JsonPath.parse(response).read("password"), aStoreAdmin.password());
+        assertEquals(JsonPath.parse(response).read("store.storeName"), aStore.name());
+        assertEquals(JsonPath.parse(response).read("store.storeAddress"), aStore.address());
+        assertEquals(JsonPath.parse(response).read("store.storeCategories"),storeTestHelper.storeCategoriesToString(aStore.storeCategories()));
+        assertEquals(JsonPath.parse(response).read("store.deliveryDistanceInKm"), aStore.deliveryDistanceInKm());
+        assertEquals(JsonPath.parse(response).read("store.availablePaymentMethods"), aStore.availablePaymentMethods());
+        assertEquals(JsonPath.parse(response).read("store.storeSchedule.openingTime"), aStore.storeSchedule().openingTime().toString());
+        assertEquals(JsonPath.parse(response).read("store.storeSchedule.closingTime"), aStore.storeSchedule().closingTime().toString());
+        assertEquals(JsonPath.parse(response).read("store.storeSchedule.openingDays"), storeTestHelper.storeOpeningDaysToString(aStore));
     }
 }
